Move public&private location into location service
