+++ src/components/location/dto/update-location.dto.ts
@@ -2,7 +2,13 @@ import { Field, InputType, ObjectType } from '@nestjs/graphql';
 import { Type } from 'class-transformer';
 import { ValidateNested } from 'class-validator';
 import { IdField, NameField } from '../../../common';
-import { Country, Region, Zone } from './location.dto';
+import {
+  Country,
+  PrivateLocation,
+  PublicLocation,
+  Region,
+  Zone,
+} from './location.dto';
 
 @InputType()
 export abstract class UpdateZone {
@@ -53,6 +59,39 @@ export abstract class UpdateCountry {
 }
 
 @InputType()
+export abstract class UpdatePrivateLocation {
+  @IdField()
+  readonly id: string;
+
+  @NameField({ nullable: true })
+  readonly name?: string;
+
+  @NameField({ nullable: true })
+  readonly publicName?: string;
+}
+
+@InputType()
+export abstract class UpdatePublicLocation {
+  @IdField()
+  readonly id: string;
+
+  @IdField({ nullable: true })
+  readonly fieldRegionId?: string;
+
+  @IdField({ nullable: true })
+  readonly marketingLocationId?: string;
+
+  @IdField({ nullable: true })
+  readonly privateLocationId?: string;
+
+  @IdField({ nullable: true })
+  readonly registryOfGeographyId?: string;
+
+  @IdField({ nullable: true })
+  readonly fundingAccountId?: string;
+}
+
+@InputType()
 export abstract class UpdateZoneInput {
   @Field()
   @Type(() => UpdateZone)
@@ -76,6 +115,22 @@ export abstract class UpdateCountryInput {
   readonly country: UpdateCountry;
 }
 
+@InputType()
+export abstract class UpdatePrivateLocationInput {
+  @Field()
+  @Type(() => UpdatePrivateLocation)
+  @ValidateNested()
+  readonly privateLocation: UpdatePrivateLocation;
+}
+
+@InputType()
+export abstract class UpdatePublicLocationInput {
+  @Field()
+  @Type(() => UpdatePublicLocation)
+  @ValidateNested()
+  readonly publicLocation: UpdatePublicLocation;
+}
+
 @ObjectType()
 export abstract class UpdateZoneOutput {
   @Field()
@@ -92,4 +147,16 @@ export abstract class UpdateRegionOutput {
 export abstract class UpdateCountryOutput {
   @Field()
   readonly country: Country;
+}
+
+@ObjectType()
+export abstract class UpdatePrivateLocationOutput {
+  @Field()
+  readonly privateLocation: PrivateLocation;
+}
+
+@ObjectType()
+export abstract class UpdatePublicLocationOutput {
+  @Field()
+  readonly publicLocation: PublicLocation;
 }
