--- src/components/private-location/dto/index.ts
@@ -1,5 +1,0 @@
-export * from './create-private-location.dto';
-export * from './list-private-location.dto';
-export * from './private-location.dto';
-export * from './update-private-location.dto';
-export * from './private-location-type.enum';
