--- src/components/private-location/index.ts
@@ -1,3 +1,0 @@
-export * from './dto';
-export * from './private-location.resolver';
-export * from './private-location.service';
