--- src/components/private-location/private-location.module.ts
@@ -1,9 +1,0 @@
-import { Module } from '@nestjs/common';
-import { PrivateLocationResolver } from './private-location.resolver';
-import { PrivateLocationService } from './private-location.service';
-
-@Module({
-  providers: [PrivateLocationResolver, PrivateLocationService],
-  exports: [PrivateLocationService],
-})
-export class PrivateLocationModule {}
