--- src/components/public-location/dto/index.ts
@@ -1,4 +1,0 @@
-export * from './create-public-location.dto';
-export * from './list-public-location.dto';
-export * from './public-location.dto';
-export * from './update-public-location.dto';
