--- src/components/public-location/index.ts
@@ -1,3 +1,0 @@
-export * from './dto';
-export * from './public-location.resolver';
-export * from './public-location.service';
